+++ src/Ouzo/Migrations/Migration.php
@@ -1,0 +1,12 @@
+<?php
+/*
+ * Copyright (c) Ouzo contributors, http://ouzoframework.org
+ * This file is made available under the MIT License (view the LICENSE file for more information).
+ */
+
+namespace Ouzo;
+
+abstract class Migration
+{
+    public abstract function run(Db $db);
+}