+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Support for PHP 5.6 is dropped. Minimal PHP version required is 7.2.
 Enhancements:
 * [Utilities] Added `Arrays.getDuplicates()`, `Arrays.getDuplicatesAssoc()`.
 * [Utilities] Added `FluentArray.getDuplicates()`, `FluentArray.getDuplicatesAssoc()`.
+* [Core] Added database migration tools (migrate:run and migrate:generate).
 
 Release 1.7.0
 --------